requirements.txt
astroid==3.3.5
coverage==7.6.4
dill==0.3.9
executing==2.1.0
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-pylint==0.21.0
python-dotenv==1.0.1
tomlkit==0.13.2
